EEPROM Range calculation: change from Passive to 1 second
and change INPx links to not be CP. We observed on mtca01 that the IOC is now
coming up with some (not all) $(P):PICOx:EEPROM_RNGy_RD (calc record) being
UDF Invalid. This was supposed to work (it worked before) since the INPx links
were all set to CP. Perhaps some ordering issue is preventing this record from
processing correctly.